+++ header/utils.h
@@ -11,7 +11,7 @@ typedef enum status_t {
 }status_t;
 
 
-long long int pow_(int x, int power);
+double pow_(double x, int power);
 bool string_is_number(const char* string);
 bool is_bad_ptr(void* ptr);
